Move project to go modules
