+++ go.mod
@@ -1,0 +1,9 @@
+module github.com/eleksir/mdserver
+
+go 1.16
+
+require (
+	github.com/bmizerany/pat v0.0.0-20210406213842-e4b6760bdd6f
+	github.com/russross/blackfriday v1.6.0
+	gopkg.in/yaml.v2 v2.4.0
+)
